add login feature and bcrypt password

+++ routes/modules/todos.js
@@ -9,11 +9,11 @@ router.get('/new', (req, res) => {
 
 router.post('/', (req, res) => {
   const name = req.body.name
-  console.log(name)
+  const UserId = req.user.id  
   Todo.create({
     name: name,
     isDone: 0,
-    UserId: 1
+    UserId
   })
     .then(() => res.redirect('/'))
     .catch(error => console.log(error))
@@ -21,8 +21,9 @@ router.post('/', (req, res) => {
 
 //細節
 router.get('/:id', (req, res) => {
+  const UserId = req.user.id
   const id = req.params.id  
-  return Todo.findByPk(id)
+  return Todo.findOne({ where: { id, UserId }})
     //.then(todo => console.log(todo))
     .then(todo => res.render('detail', { todo: todo.toJSON() }))
     //.then(todo => res.render('detail', { todo: JSON.stringify(todo) }))
@@ -31,23 +32,26 @@ router.get('/:id', (req, res) => {
 
 //編輯
 router.get('/:id/edit', (req, res) => {  
+  const UserId = req.user.id
   const id = req.params.id
-  Todo.findOne({where:{ id }})
+  Todo.findOne({ where: { id, UserId }})
     .then(todo => res.render('edit', { todo: todo.toJSON() }))
     .catch(error => console.log(error))
 })
 
 router.put('/:id', (req, res) => {
+  const UserId = req.user.id
   const id = req.params.id
   const { name, isDone } = req.body
-  Todo.update({ name: name, isDone: isDone === 'on' },{where:{id}})
+  Todo.update({ name: name, isDone: isDone === 'on' }, { where: { id, UserId }})
   .then(() => res.redirect(`/todos/${id}`))
   .catch(error => console.log(error))
 })
 //刪除
 router.delete('/:id', (req, res) => {
+  const UserId = req.user.id
   const id = req.params.id
-  Todo.destroy({where:{id}})
+  Todo.destroy({ where: { id, UserId }})
     .then(() => res.redirect('/'))
     .catch(error => console.log(error))
 })
